backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
transformers==4.57.1
torch==2.9.0
sentencepiece==0.2.1
spacy==3.7.0
pdfplumber==0.9.0
python-docx==0.8.11
chroma-db==0.8.13
tqdm==4.67.1
python-dotenv==1.0.0
